pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-processing-filter-library"
version = "1.0.0"
authors = [{name = "Gigi"}]
description = "Image and video processing filters"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "Pillow>=8.0.0",
    "opencv-python>=4.5.0",
    "psutil>=5.8.0",
]

[project.urls]
Repository = "https://github.com/robespierreprime/image-processing-filter-library"

[project.scripts]
image-filter = "image_processing_library.cli:main"